API reference

Below, you find a listing of all public methods of this package. Any other method you might find in the source code is not intended for direct usage.

Common interface

TODO with an exemplary link to fit.

QUnfold.fit — Function

fit(m, X, y) -> FittedMethod

Return a copy of the QUnfold method m that is fitted to the data set (X, y).

QUnfold.predict — Function

predict(m, X) -> Vector{Float64}

Predict the class prevalences in the data set X with the fitted method m.

QUnfold.predict_with_background — Function

predict_with_background(m, X, X_b, α=1) -> Vector{Float64}

Predict the class prevalences in the observed data set X with the fitted method m, taking into account a background measurement X_b that is scaled by α.

Quantification / unfolding methods

Missing docstring.

Missing docstring for CC. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ACC. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PCC. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PACC. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RUN. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SVD. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HDx. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HDy. Check Documenter's build log for details.

Missing docstring.

Missing docstring for IBU. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SLD. Check Documenter's build log for details.

Feature transformations

TODO.

Missing docstring.

Missing docstring for ClassTransformer. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TreeTransformer. Check Documenter's build log for details.